remove old hook
